+++ action_server/src/sema4ai/action_server/_cli_impl.py
@@ -324,6 +324,19 @@ def _add_cloud_command(command_subparser, defaults):
     add_json_output_args(list_organizations_parser)
     add_login_args(list_organizations_parser)
     add_verbose_args(list_organizations_parser, defaults)
+    
+
+def _add_get_user_oauth_config_path_command(command_subparser, defaults):
+    from sema4ai.action_server._cli_helpers import (
+        add_json_output_args,
+    )
+
+    get_user_oauth_config_path_parser = command_subparser.add_parser(
+        "get-user-oauth-config-path",
+        help="Returns the path to user's OAuth config",
+    )
+    
+    add_json_output_args(get_user_oauth_config_path_parser)
 
 
 def _create_parser():
@@ -404,6 +417,7 @@ def _create_parser():
     )
 
     _add_cloud_command(command_subparser, defaults)
+    _add_get_user_oauth_config_path_command(command_subparser, defaults)
 
     return base_parser
 
@@ -569,12 +583,9 @@ def _main_retcode(
         return 0
 
     if args and args[0] == "get-sema4ai-oauth-config":
-        from sema4ai.action_server import _oauth_config
+        from ._get_oauth_config import handle_get_sema4ai_oauth_config_command
 
-        contents = _oauth_config.FILE_CONTENTS["oauth_config.yaml"]
-        print(contents)
-
-        return 0
+        return handle_get_sema4ai_oauth_config_command()
 
     parser = _create_parser()
     base_args: ArgumentsNamespace = parser.parse_args(args)
@@ -647,6 +658,7 @@ def _main_retcode(
                 "start",
                 "new",
                 "cloud",
+                "get-user-oauth-config-path"
             ):
                 log.critical(f"Unexpected command: {command}.")
                 return 1
@@ -660,6 +672,11 @@ def _main_retcode(
                 from ._new_project import handle_new_command
 
                 return handle_new_command(base_args)
+            
+            if command == "get-user-oauth-config-path":
+                from ._get_oauth_config import handle_get_user_oauth_config_path_command
+                
+                return handle_get_user_oauth_config_path_command(base_args)
 
             migrate_import_or_start_args = typing.cast(
                 ArgumentsNamespaceMigrateImportOrStart, base_args