+++ action_server/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- Add new command: `get-user-oauth-config-path`
+  - Returns the path to user's local OAuth config file
+  - `--json` argument can be provided to get the result in JSON format
+
 ## 0.18.0 - 2024-08-27
 
 - Add to the package metadata an `action_package_version` field and rename `version` to `metadata_version`.